Move DataVector, TheoryVector to data_types
