+++ tests/likelihood/gauss_family/statistic/test_statistic.py
@@ -7,12 +7,13 @@ from numpy.testing import assert_allclose
 import pytest
 import sacc
 
+import firecrown.data_types
 import firecrown.likelihood.statistic as stat
 from firecrown.modeling_tools import ModelingTools
 from firecrown.parameters import ParamsMap
 
 
-VECTOR_CLASSES = (stat.TheoryVector, stat.DataVector)
+VECTOR_CLASSES = (firecrown.data_types.TheoryVector, firecrown.data_types.DataVector)
 
 
 def test_vector_create():
@@ -77,9 +78,9 @@ def test_ufunc_on_vector():
 
 
 def test_vector_residuals():
-    theory = stat.TheoryVector.from_list([1.0, 2.0, 3.0])
-    data = stat.DataVector.from_list([1.1, 2.1, 3.1])
-    difference = stat.residuals(data, theory)
+    theory = firecrown.data_types.TheoryVector.from_list([1.0, 2.0, 3.0])
+    data = firecrown.data_types.DataVector.from_list([1.1, 2.1, 3.1])
+    difference = firecrown.data_types.residuals(data, theory)
     assert isinstance(difference, np.ndarray)
     for cls in VECTOR_CLASSES:
         assert not isinstance(difference, cls)
